+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [1.14.0](https://gitlab.com/dd3tech/dd360/dd360-components/compare/v1.13.1...v1.14.0) (2023-01-05)
+
+
+### Features
+
+* **components:** Added Divider component ([defabd2](https://gitlab.com/dd3tech/dd360/dd360-components/commit/defabd29d30d668ad45022296de1c85219784774))
+
+## [1.13.1](https://gitlab.com/dd3tech/dd360/dd360-components/compare/v1.13.0...v1.13.1) (2023-01-05)
+
+
+### Bug Fixes
+
+* **components:** 🐛 change class of sizes to use props on the title RadioGroup ([240ef75](https://gitlab.com/dd3tech/dd360/dd360-components/commit/240ef75abdb1251dfbc490efb5da699825d00ebb))
+
+# [1.13.0](https://gitlab.com/dd3tech/dd360/dd360-components/compare/v1.12.0...v1.13.0) (2023-01-04)
+
+
+### Features
+
+* **components:** Added RadioGroup and Radio components ([f7fd78b](https://gitlab.com/dd3tech/dd360/dd360-components/commit/f7fd78b0d4d5ade57890b42fde2635790adcfc51))
+
 # [1.12.0](https://gitlab.com/dd3tech/dd360/dd360-components/compare/v1.11.0...v1.12.0) (2022-12-30)